+++ internal/credential/vault/query.go
@@ -20,13 +20,14 @@ insert into credential_vault_token (
   wt_add_seconds_to_now($7)  -- expiration_time
 );
 `
-	insertLeaseQuery = `
-insert into credential_vault_lease (
+
+	insertCredentialWithExpirationQuery = `
+insert into credential_vault_credential (
   public_id, -- $1
   library_id, -- $2
   session_id, -- $3
   token_hmac, -- $4
-  lease_id, -- $5
+  external_id, -- $5
   is_renewable, -- $6
   last_renewal_time, -- $7
   expiration_time -- $8
@@ -35,12 +36,35 @@ insert into credential_vault_lease (
   $2, -- library_id
   $3, -- session_id
   $4, -- token_hmac
-  $5, -- lease_id
+  $5, -- external_id
   $6, -- is_renewable
   $7, -- last_renewal_time
   wt_add_seconds_to_now($8)  -- expiration_time
 );
 `
+
+	insertCredentialWithInfiniteExpirationQuery = `
+insert into credential_vault_credential (
+  public_id, -- $1
+  library_id, -- $2
+  session_id, -- $3
+  token_hmac, -- $4
+  external_id, -- $5
+  is_renewable, -- $6
+  last_renewal_time, -- $7
+  expiration_time -- infinity
+) values (
+  $1, -- public_id
+  $2, -- library_id
+  $3, -- session_id
+  $4, -- token_hmac
+  $5, -- external_id
+  $6, -- is_renewable
+  $7, -- last_renewal_time
+  'infinity' -- expiration_time
+);
+`
+
 	upsertClientCertQuery = `
 insert into credential_vault_client_certificate
   (store_id, certificate, certificate_key, certificate_key_hmac, key_id)
@@ -57,5 +81,11 @@ returning *;
 	deleteClientCertQuery = `
 delete from credential_vault_client_certificate
  where store_id = $1;
+`
+
+	selectPrivateLibrariesQuery = `
+select *
+  from credential_vault_library_private
+ where public_id in (%s);
 `
 )