+++ internal/db/schema/migrations/postgres/10/01_sentinels.down.sql
@@ -1,0 +1,7 @@
+begin;
+
+  drop function wt_to_sentinel;
+  drop domain wt_sentinel;
+  drop function wt_is_sentinel;
+
+commit;